Commit to move data directory to data-structure folder, install node-postgress
maintain(data/*, models/*): move data to data-structure
- move all data files to data-structure folder
- modify path to the moved files in all dependent modules
- install node-postgres
- [Partly Delivers #165437339]
