+++ data-structure/models/user.js
@@ -1,5 +1,5 @@
 const { List, fromJS } = require('immutable');
-const userData = require('../../data/user');
+const userData = require('../data/user');
 const { asyncHashPassword } = require('../../utilities/password');
 
 const userSchema = {
