+++ package.json
@@ -31,7 +31,8 @@
     "immutable": "^4.0.0-rc.12",
     "jsonwebtoken": "^8.5.1",
     "method-override": "^3.0.0",
-    "morgan": "^1.9.1"
+    "morgan": "^1.9.1",
+    "pg": "^7.10.0"
   },
   "devDependencies": {
     "chai": "^4.2.0",
